renamer/requirements.txt
Pillow
certifi
chardet
click
configobj
Cython
dbus-python
dulwich
fastimport
Flask
idna
itsdangerous
Jinja2
MarkupSafe
Pygments
PyGObject
python-apt
PyYAML
requests
requests-unixsocket
six
urllib3
Werkzeug
